small changes in Widgets structure
