+++ Stats/Widgets/Charts/BarChart.swift
